finished refactor of place to meet initial branch goals. CREATE: PlaceUpsertRequest, ScheduleMetaDataRequest, related converters Modified : scheduleUpsertRequest -> ScheduleInsertRequest , ScheduleAccessService, ScheduleAccessController.
separated out modification of places and metadata, adhere more to SRP.
more clearly defines each action of the 'update' procedure of Schedule.

converted domain entity relation between Place and Schedule from 1 : 1 to 1 : N.
above decision was made in part to reduce the amount of duplicated Place entities should a location, such as a hotel
be repeatedly visited.

this also allows mass - changes in plans in a given itinerary to be made.

this now requires a more complex handling between places objects and schedule objects, as the events no longer cascade as of current version
in order to avoid possible unwanted side effects